Fix iterator post-increment operators.
The post-increment operators in various iterators didn't conform to the C++
standard (and, in fact, I'm not sure how they ever worked!). This required
two fixes.

First: according to section [over.inc], the post-increment operator's
parameter only exists to differentiate between pre- and post-increment,
and compliant compilers should only generate calls to it with the argument
0. Calling the method manually, e.g., `i++(42)` should still only
increment the iterator once. So, the current implementation, when called by
a compiler-generated expation of `i++`, won't even increment the iterator!

Also, post-increment operators should return a copy of the iterator from
*before* the increment was performed, not a reference to an incremented
iterator. Otherwise, `i++` would be equivalent to `++i`.