Don't abort on missing but optional values.
When constructing an AST element, it's possible for the next thing on the
stack to have the correct type but be in the wrong location (e.g., adjacent
to something of the same type). The comment at this location already
describes the issue, but the code always reports an error and returns false.
Before error'ing out in this way, we should check to see whether the AST
element we're looking for is optional: missing optional nodes aren't errors.