Add XFAIL test to demonstrate left-recursion bug.
This test demonstrates a limitation in Pegmatite's implementation of the
OMeta PEG parsing algorithm. In this test, we are able to parse a term
like `foo.bar`, and we would expect to be able to be able to parse the
term `foo.bar.baz.wibble` using the same grammar. Indeed, the OMeta
successor Ohm is able to do just that, but either Pegmatite's
implementation of the OMeta algorithm or the OMeta algorithm itself is
unable to cope with this kind of left recursion.